ims_registrar_pcscf : added a condition for subscribiton about sos param
if there is a sos parameter on Contact header on saving state in P-CSCF ,P-CSCF dont subscribe to S-CSCF for reg-events.

3GPP TS 24.229 5.1.6.3

Upon receiving the 200 (OK) to the REGISTER request that completes the emergency registration, the UE shall not subscribe to the reg event package of the public user identity specified in the REGISTER request.